library: raw injection, control transfers, rewrite rules and descriptor patches
Mirror the v3.4.0 firmware developer layer in the crate, gated on the
imperfect-clone opt-in and addressed in the CATCH (class, id, direction) space.

- Protocol version 7; HEALTH widens to a u16 with three new flags.
- New verbs: raw endpoint writes, control transfers to the real device
  (correlated on their own reply opcode by sequence), rewrite rules over every
  traffic class, and persisted descriptor patches, plus their readbacks.
- Rewrite rules join the reconnect/keepalive re-assertion as session state;
  patches persist on the box and stay out of it.
- Async wrappers, MockBox modelling, and the capi kept compiling (its own
  developer-layer surface is a follow-up).

+++ src/tests/catch.rs
@@ -555,9 +555,9 @@ fn decode_catch_through_parse_resp() {
 
 #[test]
 fn health_catch_on_bit_roundtrips() {
-    let h = Health::from_flags(H_CATCH_ON);
+    let h = Health::from_flags(u16::from(H_CATCH_ON));
     assert!(h.catch_on);
-    assert_eq!(h.to_flags(), H_CATCH_ON);
+    assert_eq!(h.to_flags(), u16::from(H_CATCH_ON));
 }
 
 #[cfg(feature = "mock")]
